Provided an alternative approach to key handling functions
Here's my alternative approach to handling the different key press keys.
In reality it's not really any shorter, but by defining the options in
an object orgranised by keyboard keys it's a little easier to read than
a long if/else statement while being easy to add to when needed (Just
add a new value to the object where the key is the `event.key` value and
the object value is the function to run).

Don't feel the need to merge this in if you do prefer the if/else way
though. Just offering the approach I'd take. Can very much be a matter
of opinion.