Made a series of changes to assist contribution
- Removed node_modules/ from the repo and added a .gitignore for common
  unwanted directories and files.
  - This is since it's not common to have these checked in to the repo
    since they'll often cause a lot of changes, which makes your own
    code harder to keep track of, while they'd likely be slightly
    different depending on operating system making cross-OS
    contributions trickier. No massive reason to include these unless
    used for deployments.
- Added a README.md file to provide guidance regarding what the project
  is and how to get started within developing on the project.
- Added http-server as a dev-dependancy, to make it easier for people to
  use the 'serve' npm script.
- Updated sass library version since I was already making changes to the
  package* files. Think this caused a change to the css sourcemaps.

Like the node_modules it is often common to not include auto generated
files such as the build css files or css sourcemaps, since their changes
can cause distruption when viewing project changes, but there could be
some use in retaining I guess.

Feel free to re-word any of the readme bits; The main focus was just to
add guidance for potential contributors.