Updated OpenDyslexic usage to use all font files
Before, only the regular font of the 4 files added were being used. This
adds support for the italic and bold combinations. In addition, I
thought I'd extract the OpenDyslexic specific rules to their own partial
to be shared across both of the accessibility theme var files for easier
management. As part of this I renamed the folder for the fonts just to
align a bit better with your folder convention.

I see that OpenDyslexic provide fonts as woff2 format also. Might be
worth using those instead if performance is a concern since woff2 are
much smaller and generally well supported these days.